requirements.txt
Flask==3.1.*
Pillow==11.*
python-dotenv==1.*
werkzeug==3.1.*
